plugins/golang-filter/semantic-router/classifier-service/requirements.txt
# gRPC相关依赖
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1

# 机器学习相关依赖
torch==2.1.0
transformers==4.35.0
tokenizers==0.14.1
numpy==1.24.3
scikit-learn==1.3.0

# 缓存和数据库
redis==5.0.1
pymongo==4.6.0

# 日志和监控
loguru==0.7.2
prometheus-client==0.19.0

# 配置管理
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# 异步支持
asyncio==3.4.3
aioredis==2.0.1

# 开发和测试
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-grpc==0.8.0
black==23.11.0
isort==5.12.0
flake8==6.1.0